+++ src/Yasumi/Provider/UnitedKingdom/England.php
@@ -17,6 +17,7 @@ declare(strict_types = 1);
 
 namespace Yasumi\Provider\UnitedKingdom;
 
+use Yasumi\Holiday;
 use Yasumi\Provider\UnitedKingdom;
 
 /**
@@ -35,4 +36,14 @@ class England extends UnitedKingdom
      * country or sub-region.
      */
     public const ID = 'GB-ENG';
+
+    /**
+     * Initialize holidays for England (United Kingdom)..
+     */
+    public function initialize(): void
+    {
+        parent::initialize();
+
+        $this->addHoliday($this->easterMonday($this->year, $this->timezone, $this->locale, Holiday::TYPE_BANK));
+    }
 }